+++ crates/assembly-syntax/src/sema/errors.rs
@@ -190,6 +190,12 @@ pub enum SemanticAnalysisError {
         #[label]
         span: SourceSpan,
     },
+    #[error("unused constant")]
+    #[diagnostic(severity(Warning), help("this constant is never used and can be safely removed"))]
+    UnusedConstant {
+        #[label]
+        span: SourceSpan,
+    },
     #[error("missing import: the referenced module has not been imported")]
     #[diagnostic()]
     MissingImport {